+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Plight of the fellowship of the ring".
 - Tests for "DDCG Mapper".
 - Tests for "Queneau numbers".
+- Tests for "Simple fraction to mixed number".
 
 ## [1.15.0] - 2022-11-30
 ### Added